Fix RestrictedHash for Rails 4.2
In 4.2 the implementation of transform_keys (that is used by
stringify_keys for example) in ActiveSupport's hash
extension changed: rails/rails@f1bad130.
Now it tries to create a new instance of the
class of the object the stringify_keys is called on and not just
creating new class. This means that in case of RestrictedHash it tries
to call RestrictedHash#new with no attributes.

Since ActiveRecord does stringify when you do Model.create(hash)
it basically makes input_sanitizer unusable with Rails 4.2.

This commit overrides the rails extension hash methods transform_keys
and transform_keys! so that it behaves correctly.

Test case: https://gist.github.com/piotrj/25106a05881ba44d38b2