Accept Procs in the `allow` option of StringCheck.
The `allow` option currently accepts an enumerable object of allowed
values. In some cases it's not enough - for example, when it's supplied
with a database query. This works fine as long as no new records are
added to a table.

Now you can provide a Proc, Lambda or any other callable object which
returns an enumerable, so it can be called on each check.